requirements.txt
Django==6.0.1
psycopg2-binary>=2.9
gunicorn>=21.2.0
whitenoise>=6.6.0
django-environ>=0.11.0
Pillow>=10.2.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
django-unfold==0.78.1
